+++ programs/dynamic-bonding-curve/src/instructions/initialize_pool/ix_initialize_virtual_pool_with_spl_token.rs
@@ -1,6 +1,7 @@
-use anchor_lang::prelude::*;
+use anchor_lang::{prelude::*, solana_program::clock::SECONDS_PER_DAY};
 use anchor_spl::{
-    token::{Mint, Token, TokenAccount},
+    token::{Mint, MintTo, Token, TokenAccount},
+    token_2022::spl_token_2022::instruction::AuthorityType,
     token_interface::{
         Mint as MintInterface, TokenAccount as TokenAccountInterface, TokenInterface,
     },
@@ -8,13 +9,20 @@ use anchor_spl::{
 use std::cmp::{max, min};
 
 use crate::{
+    activation_handler::get_current_point,
+    base_fee::BaseFeeEnumReader,
     const_pda,
-    constants::seeds::{POOL_PREFIX, TOKEN_VAULT_PREFIX},
+    constants::{
+        fee::PROTOCOL_LIQUIDITY_MIGRATION_FEE_BPS,
+        seeds::{POOL_PREFIX, TOKEN_VAULT_PREFIX},
+        MIN_LOCKED_LIQUIDITY_BPS,
+    },
+    cpi_checker::cpi_with_account_lamport_and_owner_checking,
     event::EvtInitializePool,
-    instructions::initialize_pool::process_initialize_virtual_pool_with_spl_token::process_initialize_virtual_pool_with_spl_token,
-    state::{PoolConfig, PoolType, VirtualPool},
-    token::is_supported_quote_mint,
-    PoolError,
+    process_create_token_metadata,
+    state::{fee::VolatilityTracker, BaseFeeMode, PoolConfig, PoolType, TokenType, VirtualPool},
+    token::{transfer_lamports_from_user, validate_quote_mint_with_token_badge},
+    PoolError, ProcessCreateTokenMetadataParams,
 };
 
 #[derive(AnchorSerialize, AnchorDeserialize)]
@@ -135,26 +143,121 @@ pub fn handle_initialize_virtual_pool_with_spl_token<'info>(
     ctx: Context<'info, InitializeVirtualPoolWithSplTokenCtx<'info>>,
     params: InitializePoolParameters,
 ) -> Result<()> {
+    validate_quote_mint_with_token_badge(&ctx.accounts.quote_mint, ctx.remaining_accounts.first())?;
+
+    let config = ctx.accounts.config.load()?;
+
     require!(
-        is_supported_quote_mint(&ctx.accounts.quote_mint)?,
-        PoolError::InvalidQuoteMint
+        config.get_total_liquidity_locked_bps_at_n_seconds(SECONDS_PER_DAY)?
+            >= MIN_LOCKED_LIQUIDITY_BPS,
+        PoolError::InvalidMigrationLockedLiquidity
     );
 
-    let activation_point = process_initialize_virtual_pool_with_spl_token(
-        &ctx.accounts.config,
-        &ctx.accounts.pool_authority,
-        &ctx.accounts.creator,
-        &ctx.accounts.base_mint,
-        &ctx.accounts.pool,
-        &ctx.accounts.base_vault,
-        ctx.accounts.quote_vault.key(),
-        &ctx.accounts.mint_metadata,
-        &ctx.accounts.metadata_program,
-        &ctx.accounts.payer,
-        &ctx.accounts.token_program,
-        &ctx.accounts.system_program,
-        params,
+    require!(
+        config.pool_fees.base_fee.get_base_fee_mode()? != BaseFeeMode::RateLimiter,
+        PoolError::DeprecatedBaseFeeMode
+    );
+
+    // validate min base fee
+    config.pool_fees.base_fee.validate_min_base_fee()?;
+
+    let initial_base_supply = config.get_initial_base_supply()?;
+
+    let token_type_value =
+        TokenType::try_from(config.token_type).map_err(|_| PoolError::InvalidTokenType)?;
+    require!(
+        token_type_value == TokenType::SplToken,
+        PoolError::InvalidTokenType
+    );
+
+    let InitializePoolParameters { name, symbol, uri } = params;
+
+    let token_authority = config.get_token_authority()?;
+    // mint authority option are deprecated and no longer allowed for new pools
+    require!(
+        !token_authority.has_mint_authority(),
+        PoolError::InvalidTokenAuthorityOption
+    );
+
+    // create token metadata
+    cpi_with_account_lamport_and_owner_checking(
+        || {
+            process_create_token_metadata(ProcessCreateTokenMetadataParams {
+                system_program: ctx.accounts.system_program.to_account_info(),
+                payer: ctx.accounts.payer.to_account_info(),
+                pool_authority: ctx.accounts.pool_authority.to_account_info(),
+                mint: ctx.accounts.base_mint.to_account_info(),
+                metadata_program: ctx.accounts.metadata_program.to_account_info(),
+                mint_metadata: ctx.accounts.mint_metadata.to_account_info(),
+                creator: ctx.accounts.creator.to_account_info(),
+                name: &name,
+                symbol: &symbol,
+                uri: &uri,
+                pool_authority_bump: const_pda::pool_authority::BUMP,
+                token_authority,
+                partner: config.fee_claimer,
+            })
+        },
+        ctx.accounts.pool_authority.to_account_info(),
     )?;
+
+    // mint token
+    let seeds = pool_authority_seeds!(const_pda::pool_authority::BUMP);
+    anchor_spl::token::mint_to(
+        CpiContext::new_with_signer(
+            ctx.accounts.token_program.key(),
+            MintTo {
+                mint: ctx.accounts.base_mint.to_account_info(),
+                to: ctx.accounts.base_vault.to_account_info(),
+                authority: ctx.accounts.pool_authority.to_account_info(),
+            },
+            &[&seeds[..]],
+        ),
+        initial_base_supply,
+    )?;
+
+    // revoke mint authority
+    anchor_spl::token_interface::set_authority(
+        CpiContext::new_with_signer(
+            ctx.accounts.token_program.key(),
+            anchor_spl::token_interface::SetAuthority {
+                current_authority: ctx.accounts.pool_authority.to_account_info(),
+                account_or_mint: ctx.accounts.base_mint.to_account_info(),
+            },
+            &[&seeds[..]],
+        ),
+        AuthorityType::MintTokens,
+        None,
+    )?;
+
+    // charge pool creation fee
+    if config.pool_creation_fee > 0 {
+        transfer_lamports_from_user(
+            ctx.accounts.payer.to_account_info(),
+            ctx.accounts.pool.to_account_info(),
+            ctx.accounts.system_program.to_account_info(),
+            config.pool_creation_fee,
+        )?;
+    }
+
+    // init pool
+    let mut pool = ctx.accounts.pool.load_init()?;
+
+    let activation_point = get_current_point(config.activation_type)?;
+
+    pool.initialize(
+        VolatilityTracker::default(),
+        ctx.accounts.config.key(),
+        ctx.accounts.creator.key(),
+        ctx.accounts.base_mint.key(),
+        ctx.accounts.base_vault.key(),
+        ctx.accounts.quote_vault.key(),
+        config.sqrt_start_price,
+        PoolType::SplToken.into(),
+        activation_point,
+        initial_base_supply,
+        PROTOCOL_LIQUIDITY_MIGRATION_FEE_BPS,
+    );
 
     emit_cpi!(EvtInitializePool {
         pool: ctx.accounts.pool.key(),