+++ programs/dynamic-bonding-curve/src/utils/token.rs
@@ -169,18 +169,6 @@ pub fn transfer_token_from_pool_authority<'info>(
     Ok(())
 }
 
-fn is_supported_quote_mint_extensions(
-    mint: &StateWithExtensions<spl_token_2022::state::Mint>,
-) -> Result<bool> {
-    let extensions = mint.get_extension_types()?;
-    for e in extensions {
-        if e != ExtensionType::MetadataPointer && e != ExtensionType::TokenMetadata {
-            return Ok(false);
-        }
-    }
-    Ok(true)
-}
-
 fn is_transfer_fee_zero(mint: &StateWithExtensions<spl_token_2022::state::Mint>) -> Result<bool> {
     if let Ok(transfer_fee_config) = mint.get_extension::<TransferFeeConfig>() {
         let older_transfer_fee_bps = u16::from(
@@ -215,36 +203,17 @@ pub fn validate_transfer_fee_is_zero(mint_account_info: &AccountInfo) -> Result<
 }
 
 /// Rule: quote mint must be SPL-Token, or Token-2022 (non-native) with only metadata extensions
+/// never allow a non-zero transfer fee
 pub fn is_supported_quote_mint(mint_account: &InterfaceAccount<Mint>) -> Result<bool> {
     let mint_info = mint_account.to_account_info();
     if *mint_info.owner == Token::id() {
         return Ok(true);
     }
 
-    if spl_token_2022::native_mint::check_id(&mint_account.key()) {
-        return Err(PoolError::UnsupportNativeMintToken2022.into());
-    }
-
-    let mint_data = mint_info.try_borrow_data()?;
-    let mint = StateWithExtensions::<spl_token_2022::state::Mint>::unpack(&mint_data)?;
-    is_supported_quote_mint_extensions(&mint)
-}
-
-/// Rule: same as [`is_supported_quote_mint`]
-/// except a mint with other extensions is allowed with a valid token badge
-/// never allow a non-zero transfer fee regardless of token badge
-pub fn validate_quote_mint_with_token_badge<'info>(
-    quote_mint: &InterfaceAccount<'info, Mint>,
-    token_badge: &Option<AccountLoader<'info, TokenBadge>>,
-) -> Result<()> {
-    let mint_info = quote_mint.to_account_info();
-    if mint_info.owner.eq(&Token::id()) {
-        return Ok(());
-    }
-
-    if spl_token_2022::native_mint::check_id(&quote_mint.key()) {
-        return Err(PoolError::UnsupportNativeMintToken2022.into());
-    }
+    require!(
+        !spl_token_2022::native_mint::check_id(&mint_account.key()),
+        PoolError::UnsupportNativeMintToken2022
+    );
 
     let mint_data = mint_info.try_borrow_data()?;
     let mint = StateWithExtensions::<spl_token_2022::state::Mint>::unpack(&mint_data)?;
@@ -254,10 +223,21 @@ pub fn validate_quote_mint_with_token_badge<'info>(
         PoolError::QuoteMintHasNonZeroTransferFee
     );
 
-    if !is_supported_quote_mint_extensions(&mint)? {
-        let token_badge = token_badge
-            .as_ref()
-            .ok_or_else(|| PoolError::InvalidTokenBadge)?;
+    let extensions = mint.get_extension_types()?;
+    for e in extensions {
+        if e != ExtensionType::MetadataPointer && e != ExtensionType::TokenMetadata {
+            return Ok(false);
+        }
+    }
+    Ok(true)
+}
+
+pub fn validate_quote_mint_with_token_badge<'info>(
+    quote_mint: &InterfaceAccount<'info, Mint>,
+    token_badge: Option<&'info AccountInfo<'info>>,
+) -> Result<()> {
+    if !is_supported_quote_mint(quote_mint)? {
+        let token_badge = token_badge.ok_or_else(|| PoolError::InvalidTokenBadge)?;
         require!(
             is_token_badge_initialized(quote_mint.key(), token_badge)?,
             PoolError::InvalidTokenBadge
@@ -268,8 +248,9 @@ pub fn validate_quote_mint_with_token_badge<'info>(
 
 fn is_token_badge_initialized<'info>(
     mint: Pubkey,
-    token_badge: &AccountLoader<'info, TokenBadge>,
+    token_badge: &'info AccountInfo<'info>,
 ) -> Result<bool> {
+    let token_badge: AccountLoader<'_, TokenBadge> = AccountLoader::try_from(token_badge)?;
     let token_badge = token_badge.load()?;
     Ok(token_badge.token_mint == mint)
 }